+++ src/main/java/com/gurpreet/starter/autoconfigure/StarterAutoConfiguration.java
@@ -1,10 +1,13 @@
 package com.gurpreet.starter.autoconfigure;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.gurpreet.starter.config.CorsConfig;
 import com.gurpreet.starter.config.OpenApiConfig;
 import com.gurpreet.starter.config.SecurityConfig;
 import com.gurpreet.starter.exception.GlobalExceptionHandler;
 import com.gurpreet.starter.logging.RequestLoggingFilter;
+import com.gurpreet.starter.ratelimit.RateLimitFilter;
+
 import com.gurpreet.starter.security.JwtTokenProvider;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
@@ -30,7 +33,8 @@ import org.springframework.context.annotation.Import;
         SecurityStarterProperties.class,
         CorsStarterProperties.class,
         LoggingProperties.class,
-        SwaggerProperties.class
+        SwaggerProperties.class,
+        RateLimitProperties.class
 })
 @Import({
         CorsConfig.class,
@@ -46,5 +50,12 @@ public class StarterAutoConfiguration {
     @ConditionalOnProperty(prefix = "gurpreet.starter.logging", name = "enabled", havingValue = "true", matchIfMissing = true)
     public RequestLoggingFilter requestLoggingFilter(LoggingProperties loggingProperties) {
         return new RequestLoggingFilter(loggingProperties);
+    }
+    
+    @Bean
+    @ConditionalOnMissingBean(RateLimitFilter.class)
+    @ConditionalOnProperty(prefix = "gurpreet.starter.rate-limit", name = "enabled", havingValue = "true", matchIfMissing = true)
+    public RateLimitFilter rateLimitFilter(RateLimitProperties rateLimitProperties, ObjectMapper objectMapper) {
+        return new RateLimitFilter(rateLimitProperties, objectMapper);
     }
 }